requirements.txt
xlwt==1.3.0
tabula-py==1.4.1
openpyxl==3.0.0
et-xmlfile==1.0.1